Remove duplicated test file
