+++ tests/unit_tests/test_security.py
@@ -6,6 +6,7 @@ from pytest import MonkeyPatch
 from mealie.core import security
 from mealie.core.config import get_app_settings
 from mealie.core.dependencies import validate_file_token
+from mealie.core.security.hasher import FakeHasher, PasslibHasher, get_hasher
 from mealie.db.db_setup import session_context
 from mealie.db.models.users.users import AuthMethod
 from mealie.schema.user.user import PrivateUser
@@ -253,3 +254,21 @@ def test_user_login_ldap_auth_method(monkeypatch: MonkeyPatch, ldap_user: Privat
     assert result.full_name == ldap_user.full_name
     assert result.admin == ldap_user.admin
     assert result.auth_method == AuthMethod.LDAP
+
+
+def test_get_hasher(monkeypatch: MonkeyPatch):
+    hasher = get_hasher()
+
+    assert isinstance(hasher, FakeHasher)
+
+    monkeypatch.setenv("TESTING", "0")
+
+    get_hasher.cache_clear()
+    get_app_settings.cache_clear()
+
+    hasher = get_hasher()
+
+    assert isinstance(hasher, PasslibHasher)
+
+    get_app_settings.cache_clear()
+    get_hasher.cache_clear()
